requirements.txt
asgiref==3.7.2
dj-rest-auth==4.0.1
Django==4.2.2
django-axes==6.0.5
django-filter==23.2
django-ratelimit==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.3
Pillow==10.0.0
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
